scripts/pyproject.toml
[tool.poetry]
name = "electrical-demand"
version = "0.1.0"
description = ""
authors = ["Sebastián Dalceggio <sdalceggio@gmail.com>"]
readme = "README.md"
packages = [{include = "electrical_demand"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
matplotlib = "^3.6.2"
numpy = "^1.23.5"
scikit-learn = "^1.1.3"
sqlalchemy = "^1.4.44"
alembic = "^1.8.1"
requests = "^2.28.1"
psycopg2-binary = "^2.9.5"
fsspec = "^2022.11.0"
s3fs = "^2022.11.0"
pendulum = "^2.1.2"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
python-decouple = "^3.6"
streamlit = "^1.15.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.17.1"
notebook = "^6.5.2"
black = "^22.10.0"
pylint = "^2.15.6"
apache-airflow = "^2.4.3"
docker = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
